feat(logs): add base logs

+++ src/packages/core/src/services/spots/spot.ts
@@ -1,13 +1,21 @@
 import type { Spot } from '@your-spot/contracts'
 import { spotCollection } from '@your-spot/database'
 
-import { stringToObjectId } from '../common'
+import {
+    isNotValid,
+    stringToObjectId,
+} from '../common'
 
 import { createMongoCoordinate } from './coordinate'
 import { createSpot as createSpotFromDbSpot } from './mapper'
 
 
 export function createSpot(spot: Omit<Spot, 'id'>) {
+    if (spot.authorId && isNotValid(spot.authorId)) {
+        console.log('spot/createSpot Wrong authorId were passed.', spot)
+        throw new Error('AuthorId is not valid.')
+    }
+
     const authorObjectId = spot.authorId
         ? stringToObjectId(spot.authorId)
         : null
@@ -22,6 +30,11 @@ export function createSpot(spot: Omit<Spot, 'id'>) {
 }
 
 export async function deleteSpot(spotId: string) {
+    if (isNotValid(spotId)) {
+        console.log('spot/deleteSpot Wrong id were passed.', spotId)
+        throw new Error('Id is not valid.')
+    }
+
     const objectId = stringToObjectId(spotId)
 
     await spotCollection.deleteOne({
@@ -30,6 +43,11 @@ export async function deleteSpot(spotId: string) {
 }
 
 export async function updateSpot(spot: Spot) {
+    if (isNotValid(spot.id)) {
+        console.log('spot/updateSpot Wrong id were passed.', spot)
+        throw new Error('Id is not valid.')
+    }
+
     const objectId = stringToObjectId(spot.id)
 
     await spotCollection.updateOne({
@@ -45,14 +63,12 @@ export async function updateSpot(spot: Spot) {
 }
 
 export async function getSpot(spotId: string) {
-    let objectId = null
-    try {
-        objectId = stringToObjectId(spotId)
-    }
-    catch
-    {
+    if (isNotValid(spotId)) {
+        console.log('spot/getSpot Wrong id were passed.', spotId)
         return null
     }
+
+    const objectId = stringToObjectId(spotId)
 
     const dbSpot = await spotCollection.findOne({
         _id: objectId,
@@ -65,6 +81,11 @@ export async function getSpot(spotId: string) {
 }
 
 export async function getSpots(ids: ReadonlyArray<string>) {
+    if (ids.some(isNotValid)) {
+        console.warn('spot/getSpots Wrong ids were passed.', ids)
+        return []
+    }
+
     const objectIds = ids.map(stringToObjectId)
 
     const dbSpots = await spotCollection
@@ -79,14 +100,12 @@ export async function getSpots(ids: ReadonlyArray<string>) {
 }
 
 export async function getSpotsForAuthor(authorId: string) {
-    let authorObjectId = null
-    try {
-        authorObjectId = stringToObjectId(authorId)
-    }
-    catch
-    {
+    if (isNotValid(authorId)) {
+        console.warn('spot/getSpotsForAuthor Wrong id were passed.', authorId)
         return []
     }
+
+    const authorObjectId = stringToObjectId(authorId)
 
     const dbSpots = await spotCollection.find({ authorId: authorObjectId }).toArray()
